% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinsConstructor.R
\docType{class}
\name{BinNode-class}
\alias{BinNode-class}
\alias{BinNode}
\title{Reference Class to represent a bin for creating a tree of bins
In which the leaf bins are the final bins that will be used}
\description{
Reference Class to represent a bin for creating a tree of bins
In which the leaf bins are the final bins that will be used
}
\section{Fields}{

\describe{
\item{\code{range}}{A matrix which holds the start and the end values for each dimension}

\item{\code{count}}{The number of the points}

\item{\code{parent}}{The parent bin of the bin in the tree}

\item{\code{isLeftSon}}{is the son in the tree is left son}
}}


\keyword{internal}
