% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENA.R
\name{CENA}
\alias{CENA}
\title{The Cell Niche Associations (CENA) algorithm}
\usage{
CENA(geneExpressionDataMatrix, phenotypeData, cellSpace,
  resolution_parameter, genesToRun = row.names(geneExpressionDataMatrix),
  no_cores = NULL, k1 = NULL, k2 = 10, Tw = 30,
  python_path = NULL)
}
\arguments{
\item{geneExpressionDataMatrix}{A matrix containing the single-cell RNA-seq data.
Each row corresponds to a certain gene and each column to a certain cell.
The algorithm assumes the order of the cells in this scRNA-seq data matrix is the same as the order in the meta-data ('phenotypeData') and the cell-state space ('cellSpace').}

\item{phenotypeData}{A vector containing the meta-data levels of each of the cells.}

\item{cellSpace}{Cell coordinates matrix for the single-cell data. Each row represent a cell and each column corresponds to a specific dimension. Only 2-dim spaces are allowed.}

\item{resolution_parameter}{The resolution parameter of Leiden. Cutoff as desired; Higher values provide smaller clusters.}

\item{genesToRun}{A vector of genes for which associations should be inferred. The default value is all genes in the input scRNA-seq matrix.
However, to reduce running times, it is recommended to use a small subset of the genes.}

\item{no_cores}{The number of cores which will be used for the analysis. The defalt (NULL) is total number of cores minus 1.}

\item{k1}{The K1 parameter. The default value is 1 percentage of the cells. k1 should be kept relatively small.}

\item{k2}{The K2 paremeter. The default value is 10. k2 have relatively small effect on results.}

\item{Tw}{The subset size cutoff (Tw). The default value is 30. Cutoff as desired; high values filter out small subsets.}

\item{python_path}{The location of python on the machine. If not specified, the default python path will be selected.
Note that python3 and the following libraries should be installed: numpy, igraph, leidenalg.}
}
\value{
\describe{
 \item{cluster_information}{Association statistics for cell subsets obtained by CENA for each of the genes (rows).
 Columns corresponds to the r.squared, slope, intercept and p.value of these associations.}
 \item{cluster}{A list containing, for each gene, the identified cell subset in which the association occurs.}
 \item{parameters}{The initial parameters used by CENA.}
}
}
\description{
CENA is a method for a joint identification of pairwise association together with
the particular subset of cells in which the association is detected.
In this implementation, CENA is limited to associations between the genes' expression levels
(data from scRNA-sequencing) and an additional cellular meta-data of choice.
Note that python3 should be installed on the machine, with the following libraries: numpy, igraph, leidenalg.
}
\examples{
data(cellSpace)
data(geneExpressionDataMatrix)
data(phenotypeData)
# running CENA on 5 genes
results = CENA(geneExpressionDataMatrix, phenotypeData, cellSpace, resolution_parameter = 8, no_cores = 1)
}
