% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENA.R
\name{specificity}
\alias{specificity}
\title{Specificity check for CENA predictions}
\usage{
specificity(prediction, geneExpressionDataMatrix, phenotypeData, cellSpace,
  genesToRun, no_cores = NULL, numberOfRepeats = 100)
}
\arguments{
\item{prediction}{The result of the CENA run.}

\item{geneExpressionDataMatrix}{A matrix containing the single-cell RNA-seq data.
Each row corresponds to a certain gene and each column to a certain cell.
The algorithm assumes the order of the cells in this scRNA-seq data matrix is the same as the order in the meta-data ('phenotypeData') and the cell-state space ('cellSpace').}

\item{phenotypeData}{A vector containing the meta-data levels of each of the cells.}

\item{cellSpace}{Cell coordinates matrix for the single-cell data. Each row represent a cell and each column corresponds to a specific dimension. Only 2-dim spaces are allowed.}

\item{genesToRun}{A vector of genes for which specificity will be calculated. Due to relatively long running times, it is advised to choose only genes with informative clusters inferred by CENA.}

\item{no_cores}{The number of cores which will be used for the analysis. The defalt (NULL) is total number of cores minus 1.}

\item{numberOfRepeats}{Number of CENA runs used to calculate cluster specificity. The default value is 100.}
}
\value{
A vector of p-values (specificity scores) for each of the genes.
}
\description{
Calculate the specificity scores for selected genes.
}
\examples{
data(cellSpace)
data(geneExpressionDataMatrix)
data(phenotypeData)
# running CENA on 5 genes
results = CENA(geneExpressionDataMatrix, phenotypeData, cellSpace, resolution_parameter = 8, no_cores = 1)
genesWithClusters = row.names(geneExpressionDataMatrix)[which(!is.na(rowSums(results$cluster_information)))]
specificityResults = specificity(results, geneExpressionDataMatrix, phenotypeData, cellSpace, genesToRun = genesWithClusters, no_cores = NULL, numberOfRepeats =100)
}
