% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENA.R
\name{uniqueness}
\alias{uniqueness}
\title{Uniqueness check for CENA predictions}
\usage{
uniqueness(prediction, geneExpressionDataMatrix, phenotypeData, cellSpace,
  genesToRun = names(prediction$cluster), numberOfRepeats = 1000)
}
\arguments{
\item{prediction}{The result of the CENA run.}

\item{geneExpressionDataMatrix}{A matrix containing the single-cell RNA-seq data.
Each row corresponds to a certain gene and each column to a certain cell.
The algorithm assumes the order of the cells in this scRNA-seq data matrix is the same as the order in the meta-data feature (‘phenotypeData’) and the cell-state space (‘cellSpace’) parameters.}

\item{phenotypeData}{A vector containing the meta-data levels of each of the cells.}

\item{cellSpace}{The cell space corresponding to the single-cell data. It should be a matrix for a 2 dimensional space where each column represents a dimension and each row represents a cell.}

\item{genesToRun}{A vector of genes for which robustness will be calculated. Due to relatively long running times, it is advised to choose only genes with informative clusters inferred by CENA.}

\item{numberOfRepeats}{Number of sampled clusters to compare the real r-square to. The default value is 1000.}
}
\value{
A vector of uniqueness p-values of each gene.
}
\description{
The uniqueness function takes CENA results, and for each gene,samples subset of cells multiple times in the size of the original cluster and compares
the r-square of CENA to the r-squares of the sampled clusters.
}
\examples{
data(cellSpace)
data(geneExpressionDataMatrix)
data(phenotypeData)
# running CENA on 5 genes
results = CENA(geneExpressionDataMatrix, phenotypeData, cellSpace, resolution_parameter = 8, no_cores = 1)
uniquenessResults = uniqueness(results, geneExpressionDataMatrix, phenotypeData, cellSpace, genesToRun = row.names(geneExpressionDataMatrix), numberOfRepeats =1000)
}
