% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENA.R
\name{robustness}
\alias{robustness}
\title{The Cell Niche Associations (CENA) algorithm}
\usage{
robustness(prediction, geneExpressionDataMatrix, phenotypeData, cellSpace,
  resolution_parameter, genesToRun = row.names(geneExpressionDataMatrix),
  numberOfRepeats = 100, minCoverage = 0.5, ...)
}
\arguments{
\item{prediction}{The result of the original CENA run.}

\item{geneExpressionDataMatrix}{A matrix containing the single-cell RNA-seq data.
Each row corresponds to a certain gene and each column to a certain cell.
The algorithm assumes the order of the cells in this scRNA-seq data matrix is the same as the order in the meta-data feature (‘phenotypeData’) and the cell-state space (‘cellSpace’) parameters.}

\item{phenotypeData}{A vector containing the meta-data levels of each of the cells.}

\item{cellSpace}{The cell space corresponding to the single-cell data. It should be a matrix for a 2 dimensional space where each column represents a dimension and each row represents a cell.}

\item{resolution_parameter}{the resolution of the Leiden algorithm.}

\item{genesToRun}{A vector of genes for robustness will be calculated. Due to relatively long running times, it is advised to choose only genes with informative clusters inferred by CENA.}

\item{numberOfRepeats}{Number of CENA runs used to calculate cluster robustness. The default value is 100.}

\item{minCoverage}{The minimum coverage required for matching between the predicted clusters and the clusters in each iteration. The default value is 0.5.}

\item{...}{All other parameters used by CENA.}
}
\value{
Predictions for each gene-vs-metadata analysis.
For each such pair, the prediction includes details about the association (here, ‘cluster_information’) as well as details about the cell subset of the association (here, termed ‘cluster’).
\describe{
\itemize{ A matrix where rows are genes and columns are different robustness measures.
 \item{Repetitivity}{A fraction representing how much the predicted gene clusters are shared across iterations. }
 \item{Average rank mean}{ Average cluster scores. Values range between 0 to 1, where a lower value means the perdicted gene cluster is more likely to be selected across iterations. }
 \item{Average rank Sd}{ The standard variation of cluster scores across iterations. }
 }
}
}
\description{
CENA is a method for a joint identification of pairwise association together with
the particular subset of cells in which the association is detected.
CENA does not rely on predefined cell subset but only requires that cells in the
identified cell subset would have a similar cell state. The algorithm relies on the
input cell-state space to ensure a common cell state of all cells in the inferred cell subset.
In this implementation, CENA tests association between multiple pairs of features:
one of these features is the expression of a gene (data from scRNA-sequencing)
and one additional feature is meta-data about each cell.
The algorithm can run on a list of genes, associating each of these gene with the same meta-data feature.
Note that python3 should be installed on the machine, with the following libraries: numpy, igraph, leidenalg.
}
\examples{
data(cellSpace)
data(geneExpressionDataMatrix)
data(phenotypeData)
# running CENA on 5 genes
results = CENA(geneExpressionDataMatrix, phenotypeData, cellSpace, resolution_parameter = 8, no_cores = 1)
robustnessResults = robustness(results, geneExpressionDataMatrix, phenotypeData, cellSpace, resolution_parameter = 8, no_cores = 1)
}
